Assn 14 plan added

+++ cs1400/Assn14/assn14-task1.py
@@ -3,11 +3,11 @@ from cuboid import Cuboid
 
 def main():
     dimensions = input("Please enter cube1's dimensions (ex: length, width, height): ")
-    length, width, height = [int(dimension) for dimension in dimensions.split(',')]
+    length, width, height = [float(dimension) for dimension in dimensions.split(',')]
     cube1 = Cuboid(length, width, height)
 
     dimensions = input("Please enter cube2's dimensions (ex: length, width, height): ")
-    length, width, height = [int(dimension) for dimension in dimensions.split(',')]
+    length, width, height = [float(dimension) for dimension in dimensions.split(',')]
 
     cube2 = Cuboid(length, width, height)
